+++ CodeRed/DirectX12/DirectX12GraphicsCommandList.hpp
@@ -77,6 +77,10 @@ namespace CodeRed {
 			const ResourceLayout old_layout, 
 			const ResourceLayout new_layout) override;
 
+		void resolveTexture(
+			const TextureResolveInfo& source, 
+			const TextureResolveInfo& destination) override;
+		
 		void copyBuffer(
 			const std::shared_ptr<GpuBuffer>& source, 
 			const std::shared_ptr<GpuBuffer>& destination,